lb: moving ring hash to extensions
