+++ changelogs/current.yaml
@@ -4,7 +4,7 @@ behavior_changes:
 # *Changes that are expected to cause an incompatibility if applicable; deployment changes are likely required*
 - area: build
   change: |
-    Moved the subset and maglev LB code into extensions. If you use these load balancers and override
+    Moved the subset, ring_hash, and maglev LB code into extensions. If you use these load balancers and override
     extensions_build_config.bzl you will need to include them explicitly.
 
 minor_behavior_changes:
