+++ source/common/upstream/BUILD
@@ -85,7 +85,6 @@ envoy_cc_library(
         ":load_balancer_lib",
         ":load_stats_reporter_lib",
         ":od_cds_api_lib",
-        ":ring_hash_lb_lib",
         ":host_utility_lib",
         "//envoy/api:api_interface",
         "//envoy/config:xds_resources_delegate_interface",
@@ -380,21 +379,6 @@ envoy_cc_library(
         "//source/common/config:metadata_lib",
         "//source/common/config:well_known_names",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_library(
-    name = "ring_hash_lb_lib",
-    srcs = ["ring_hash_lb.cc"],
-    hdrs = ["ring_hash_lb.h"],
-    external_deps = [
-        "abseil_inlined_vector",
-    ],
-    deps = [
-        ":thread_aware_lb_lib",
-        "//source/common/common:minimal_logger_lib",
-        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/load_balancing_policies/ring_hash/v3:pkg_cc_proto",
     ],
 )
