+++ source/extensions/load_balancing_policies/subset/BUILD
@@ -22,9 +22,9 @@ envoy_cc_library(
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
         "//source/common/upstream:load_balancer_lib",
-        "//source/common/upstream:ring_hash_lb_lib",
         "//source/common/upstream:upstream_lib",
         "//source/extensions/load_balancing_policies/maglev:maglev_lb_lib",
+        "//source/extensions/load_balancing_policies/ring_hash:ring_hash_lb_lib",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
     ],
