+++ tools/code_format/config.yaml
@@ -320,3 +320,4 @@ visibility_excludes:
 - source/extensions/config_subscription/rest/BUILD
 - source/extensions/config_subscription/filesystem/BUILD
 - source/extensions/load_balancing_policies/subset/BUILD
+- source/extensions/load_balancing_policies/ring_hash/BUILD
